+++ src/Entity/Activity.php
@@ -69,6 +69,7 @@ class Activity
     // keep the trait include exactly here, for placing the column at the correct position
     use RatesTrait;
     use ColorTrait;
+    use BudgetTrait;
 
     /**
      * @return int
