+++ src/Entity/Customer.php
@@ -150,6 +150,7 @@ class Customer
     // keep the trait include exactly here, for placing the column at the correct position
     use RatesTrait;
     use ColorTrait;
+    use BudgetTrait;
 
     /**
      * @return int
