+++ src/Form/CustomerEditForm.php
@@ -10,10 +10,6 @@
 namespace App\Form;
 
 use App\Entity\Customer;
-use App\Form\Type\ColorPickerType;
-use App\Form\Type\FixedRateType;
-use App\Form\Type\HourlyRateType;
-use App\Form\Type\YesNoType;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\CountryType;
 use Symfony\Component\Form\Extension\Core\Type\CurrencyType;
@@ -26,11 +22,10 @@ use Symfony\Component\Form\Extension\Core\Type\UrlType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-/**
- * Defines the form used to edit Customer entities.
- */
 class CustomerEditForm extends AbstractType
 {
+    use EntityFormTrait;
+
     /**
      * {@inheritdoc}
      */
@@ -101,18 +96,9 @@ class CustomerEditForm extends AbstractType
             ])
             ->add('timezone', TimezoneType::class, [
                 'label' => 'label.timezone',
-            ])
-            ->add('color', ColorPickerType::class)
-            ->add('fixedRate', FixedRateType::class, [
-                'currency' => $currency ?? false,
-            ])
-            ->add('hourlyRate', HourlyRateType::class, [
-                'currency' => $currency ?? false,
-            ])
-            ->add('visible', YesNoType::class, [
-                'label' => 'label.visible',
-            ])
-        ;
+            ]);
+
+        $this->addCommonFields($builder, $currency);
     }
 
     /**
